test: move out tests for delegation to the diiferent test file & remove duplicated tests
